Make changesPlugin.vim aware of VCS
This can be used, to have changes.vim indicate the changes
against a checked in version of a file. Currently, this
supports git, bazaar, svn, mercurial, svk and cvs.

This needs to be tested to work correctly. So use with care!
I have only tested, it works with git.

Additionally, the Makefile was broken. So fix it, that it will correctly
increment the version numbers, the date and the GLVS header.