Use a location list instead of a qf-list
The overview over which lines changed, are now located into a
locationlist. The advantage is, that an existing quickfix list won't be
destroyed.
Using bazaar as VCS has been tested and should work now.
Left todo:
There is still a redrawing problem. I need to find a better way to
display the messages.